Split createPages
